requirements.txt

streamlit
spotipy
yt_dlp
pydub
librosa
faiss-cpu
